Update dependencies
As noted in /Cargo.toml, pest is still keeping
thiserror behind but a commit on master has
addressed this so I just need to wait for the next
release.

mlua is using proc-macro-error, which is
unmaintained according to RUSTSEC-2024-0370.
I've opened an issue and PR on mlua to switch
it to proc-macro-error2.